Try to stop user from receiving notifications of users previously logged in on same device

+++ app/src/main/java/com/example/virtualbreak/view/view_fragments/logout/LogoutFragment.kt
@@ -2,6 +2,7 @@ package com.example.virtualbreak.view.view_fragments.logout
 
 import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -12,10 +13,13 @@ import com.example.virtualbreak.controller.communication.PushData
 import com.example.virtualbreak.model.Status
 import com.example.virtualbreak.view.view_activitys.MainActivity
 import com.example.virtualbreak.view.view_activitys.breakroom.BreakroomWidgetService
+import com.google.android.gms.auth.api.signin.internal.Storage
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.installations.FirebaseInstallations
 import com.google.firebase.ktx.Firebase
 import com.google.firebase.messaging.ktx.messaging
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 
 class LogoutFragment : Fragment() {
 
@@ -31,18 +35,23 @@ class LogoutFragment : Fragment() {
             PushData.setStatus(Status.ABSENT)
             //close Widget
             activity?.stopService(Intent(activity, BreakroomWidgetService::class.java))
-            //Sign out
-            Firebase.auth.signOut()
             //Unregister from fcm push notifications
             Firebase.messaging.isAutoInitEnabled = false
-            Firebase.messaging.deleteToken()
-            FirebaseInstallations.getInstance().delete()
-            startActivity(
-                Intent(
-                    activity,
-                    MainActivity::class.java
-                )
-            ) //go back to sign up/login activity
+            Firebase.messaging.deleteToken().addOnCompleteListener{ task ->
+                if (task.isSuccessful) {
+                    Log.d("LogoutFragment", "Token successfully deleted")
+                    //Sign out
+                    Firebase.auth.signOut()
+                    Log.d("LogoutFragment", "Signed out")
+                    GlobalScope.launch {
+                        FirebaseInstallations.getInstance().delete()
+                    }
+                    startActivity(Intent(activity, MainActivity::class.java)) // go back to sign up/login activity
+                } else {
+                    Log.d("LogoutFragment", "Problem deleting token")
+                    // do nothing
+                }
+            }
         }
 
         return root
